requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
exceptiongroup==1.3.1
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
lxml==6.0.2
MarkupSafe==3.0.3
openai==2.21.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3

